analise-geo-vendas/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
pydeck>=0.8.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
geopy>=2.4.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
